+++ test/OpenTelemetry.Instrumentation.StackExchangeRedis.Tests/StackExchangeRedisCallsInstrumentationTests.cs
@@ -44,6 +44,60 @@ namespace OpenTelemetry.Instrumentation.StackExchangeRedis.Tests
         [Trait("CategoryName", "RedisIntegrationTests")]
         [SkipUnlessEnvVarFoundTheory(RedisEndPointEnvVarName)]
         [InlineData("value1")]
+        public void SuccessfulCommandTestWithKey(string value)
+        {
+            var connectionOptions = new ConfigurationOptions
+            {
+                AbortOnConnectFail = true,
+            };
+            connectionOptions.EndPoints.Add(RedisEndPoint);
+
+            using var connection = ConnectionMultiplexer.Connect(connectionOptions);
+            var db = connection.GetDatabase();
+            db.KeyDelete("key1");
+
+            var activityProcessor = new Mock<BaseProcessor<Activity>>();
+            var sampler = new TestSampler();
+            using (Sdk.CreateTracerProviderBuilder()
+                .AddProcessor(activityProcessor.Object)
+                .SetSampler(sampler)
+                .AddRedisInstrumentation(connection, c => c.SetVerboseDatabaseStatements = true)
+                .Build())
+            {
+                var prepared = LuaScript.Prepare("redis.call('set', @key, @value)");
+                db.ScriptEvaluate(prepared, new { key = (RedisKey)"mykey", value = 123 });
+
+                var redisValue = db.StringGet("key1");
+
+                Assert.False(redisValue.HasValue);
+
+                bool set = db.StringSet("key1", value, TimeSpan.FromSeconds(60));
+
+                Assert.True(set);
+
+                redisValue = db.StringGet("key1");
+
+                Assert.True(redisValue.HasValue);
+                Assert.Equal(value, redisValue.ToString());
+            }
+
+            // Disposing SDK should flush the Redis profiling session immediately.
+
+            Assert.Equal(11, activityProcessor.Invocations.Count);
+
+            var scriptActivity = (Activity)activityProcessor.Invocations[1].Arguments[0];
+            Assert.Equal("EVAL", scriptActivity.DisplayName);
+            Assert.Equal("EVAL redis.call('set', ARGV[1], ARGV[2])", scriptActivity.GetTagValue(SemanticConventions.AttributeDbStatement));
+
+            VerifyActivityData((Activity)activityProcessor.Invocations[3].Arguments[0], false, connection.GetEndPoints()[0], true);
+            VerifyActivityData((Activity)activityProcessor.Invocations[5].Arguments[0], true, connection.GetEndPoints()[0], true);
+            VerifyActivityData((Activity)activityProcessor.Invocations[7].Arguments[0], false, connection.GetEndPoints()[0], true);
+            VerifySamplingParameters(sampler.LatestSamplingParameters);
+        }
+
+        [Trait("CategoryName", "RedisIntegrationTests")]
+        [SkipUnlessEnvVarFoundTheory(RedisEndPointEnvVarName)]
+        [InlineData("value1")]
         public void SuccessfulCommandTest(string value)
         {
             var connectionOptions = new ConfigurationOptions
@@ -59,7 +113,7 @@ namespace OpenTelemetry.Instrumentation.StackExchangeRedis.Tests
             using (Sdk.CreateTracerProviderBuilder()
                 .AddProcessor(activityProcessor.Object)
                 .SetSampler(sampler)
-                .AddRedisInstrumentation(connection)
+                .AddRedisInstrumentation(connection, c => c.SetVerboseDatabaseStatements = false)
                 .Build())
             {
                 var db = connection.GetDatabase();
@@ -78,8 +132,8 @@ namespace OpenTelemetry.Instrumentation.StackExchangeRedis.Tests
 
             Assert.Equal(7, activityProcessor.Invocations.Count);
 
-            VerifyActivityData((Activity)activityProcessor.Invocations[1].Arguments[0], true, connection.GetEndPoints()[0]);
-            VerifyActivityData((Activity)activityProcessor.Invocations[3].Arguments[0], false, connection.GetEndPoints()[0]);
+            VerifyActivityData((Activity)activityProcessor.Invocations[1].Arguments[0], true, connection.GetEndPoints()[0], false);
+            VerifyActivityData((Activity)activityProcessor.Invocations[3].Arguments[0], false, connection.GetEndPoints()[0], false);
             VerifySamplingParameters(sampler.LatestSamplingParameters);
         }
 
@@ -102,6 +156,55 @@ namespace OpenTelemetry.Instrumentation.StackExchangeRedis.Tests
             await Task.Delay(1).ContinueWith((t) => { third = profilerFactory(); });
             Assert.Equal(first, second);
             Assert.Equal(second, third);
+        }
+
+        [Trait("CategoryName", "RedisIntegrationTests")]
+        [SkipUnlessEnvVarFoundTheory(RedisEndPointEnvVarName)]
+        [InlineData("value1")]
+        public void CanEnrichActivityFromCommand(string value)
+        {
+            var connectionOptions = new ConfigurationOptions
+            {
+                AbortOnConnectFail = true,
+            };
+            connectionOptions.EndPoints.Add(RedisEndPoint);
+
+            using var connection = ConnectionMultiplexer.Connect(connectionOptions);
+
+            var activityProcessor = new Mock<BaseProcessor<Activity>>();
+            var sampler = new TestSampler();
+            using (Sdk.CreateTracerProviderBuilder()
+                .AddProcessor(activityProcessor.Object)
+                .SetSampler(sampler)
+                .AddRedisInstrumentation(connection, c => c.Enrich = (activity, command) =>
+                {
+                    if (command.ElapsedTime < TimeSpan.FromMilliseconds(100))
+                    {
+                        activity.AddTag("is_fast", true);
+                    }
+                })
+                .Build())
+            {
+                var db = connection.GetDatabase();
+
+                bool set = db.StringSet("key1", value, TimeSpan.FromSeconds(60));
+
+                Assert.True(set);
+
+                var redisValue = db.StringGet("key1");
+
+                Assert.True(redisValue.HasValue);
+                Assert.Equal(value, redisValue.ToString());
+            }
+
+            // Disposing SDK should flush the Redis profiling session immediately.
+
+            Assert.Equal(7, activityProcessor.Invocations.Count);
+
+            var setActivity = (Activity)activityProcessor.Invocations[1].Arguments[0];
+            Assert.Equal(true, setActivity.GetTagValue("is_fast"));
+            var getActivity = (Activity)activityProcessor.Invocations[3].Arguments[0];
+            Assert.Equal(true, getActivity.GetTagValue("is_fast"));
         }
 
         [Fact]
@@ -255,17 +358,31 @@ namespace OpenTelemetry.Instrumentation.StackExchangeRedis.Tests
             Assert.Throws<InvalidOperationException>(() => serviceProvider.GetRequiredService<TracerProvider>());
         }
 
-        private static void VerifyActivityData(Activity activity, bool isSet, EndPoint endPoint)
+        private static void VerifyActivityData(Activity activity, bool isSet, EndPoint endPoint, bool setCommandKey = false)
         {
             if (isSet)
             {
                 Assert.Equal("SETEX", activity.DisplayName);
-                Assert.Equal("SETEX", activity.GetTagValue(SemanticConventions.AttributeDbStatement));
+                if (setCommandKey)
+                {
+                    Assert.Equal("SETEX key1", activity.GetTagValue(SemanticConventions.AttributeDbStatement));
+                }
+                else
+                {
+                    Assert.Equal("SETEX", activity.GetTagValue(SemanticConventions.AttributeDbStatement));
+                }
             }
             else
             {
                 Assert.Equal("GET", activity.DisplayName);
-                Assert.Equal("GET", activity.GetTagValue(SemanticConventions.AttributeDbStatement));
+                if (setCommandKey)
+                {
+                    Assert.Equal("GET key1", activity.GetTagValue(SemanticConventions.AttributeDbStatement));
+                }
+                else
+                {
+                    Assert.Equal("GET", activity.GetTagValue(SemanticConventions.AttributeDbStatement));
+                }
             }
 
             Assert.Equal(Status.Unset, activity.GetStatus());